+++ gateway/src/constants.rs
@@ -1,0 +1,1 @@
+pub const DEFAULT_BOARD_SIZE: usize = 19;
