+++ gateway/src/kafka_io.rs
